requirements.txt
requests
gtfs-realtime-bindings
pillow
